news-ai-agent-updated/requirements.txt
streamlit
openai
requests
beautifulsoup4
readability-lxml
feedparser
pandas
lxml
